Move localisation helpers to shared scope
