+++ EssentialApp/EssentialAppTests/Helpers/SharedTestHelpers.swift
@@ -23,3 +23,19 @@ func anyData() -> Data {
 func uniqueFeed() -> [FeedImage] {
     return [FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())]
 }
+
+class DummyView: ResourceView {
+    func display(_ viewModel: Any) { }
+}
+
+var loadError: String {
+    LoadResourcePresenter<Any, DummyView>.loadError
+}
+
+var feedTitle: String {
+    FeedPresenter.title
+}
+
+var commentsTitle: String {
+    ImageCommentsPresenter.title
+}
